Update index.html
The inventory system has been updated. Rewards in your inventory without a timer will now show a "Redeem" button. Clicking it moves the item to a new "Used Rewards" section, accessible via an accordion at the bottom of the Inventory tab. Rewards with timers still have an "Activate" button and will be moved to the "Active Rewards" panel when used.